requirements.txt
streamlit>=1.36.0
streamlit-mic-recorder>=0.0.8
google-genai>=0.3.0
python-dotenv>=1.0.1
